Change: Organize the structure
