+++ src/cli/mod.rs
@@ -1,0 +1,8 @@
+pub mod build;
+pub mod error;
+pub mod jit;
+pub mod opts;
+pub mod shell;
+
+pub use build::build;
+pub use jit::run;
