+++ src/cli/shell.rs
