+++ src/compile/mod.rs
@@ -1,7 +1,1 @@
-pub mod build;
-pub mod jit;
 pub mod llvm;
-pub mod opts;
-
-pub use build::build;
-pub use jit::run;
