+++ src/lib.rs
@@ -1,11 +1,11 @@
 #![feature(box_syntax, box_patterns, never_type)]
 
+pub mod cli;
 pub mod compile;
 pub mod expression;
 pub mod ir;
 pub mod parser;
 pub mod scope;
-pub mod shell;
 pub mod test;
 pub mod transform;
 pub mod translator;
