+++ src/main.rs
@@ -1,7 +1,7 @@
 use ansi_term::Colour::Red;
 use structopt::StructOpt;
 
-use expressi::compile::{self, opts};
+use expressi::cli::{self, opts};
 
 #[derive(StructOpt)]
 enum Opt {
@@ -20,8 +20,8 @@ enum Opt {
 #[cfg_attr(tarpaulin, skip)]
 fn main() {
     let result = match Opt::from_args() {
-        Opt::Run { opt } => compile::run(&opt).map(|_| ()),
-        Opt::Build { opt } => compile::build(&opt),
+        Opt::Run { opt } => cli::run(&opt).map(|_| ()),
+        Opt::Build { opt } => cli::build(&opt),
     };
     if let Err(e) = result {
         eprintln!("{}: {}", Red.paint("Fatal Error"), e);
